load more pagination takes over settings
This pull request allows you to set the number of images displayed and the number of images to be reloaded in the settings. The two fields # Images and # Images Reloaded were previously ignored. In addition, the load more button is only displayed if the total number of images exceeds the value of # images.