Lightgallery caption options for title and/or description
This pull request allows you to select whether the title, description, or both are displayed in the lightgallery caption. In addition, a bug has been fixed that prevented captions from being displayed at all when an image link to the detail view was combined with a title link to the lightgallery.